+++ .github/workflows/release.yml
@@ -41,8 +41,8 @@ jobs:
           GOOS: ${{ matrix.os }}
           GOARCH: ${{ matrix.arch }}
           CGO_ENABLED: 0
-          LDFLAGS: -s -w -X github.com/tarampampam/mikrotik-hosts-parser/version.version=${{ steps.values.outputs.version }}
-        run: go build -trimpath -ldflags "$LDFLAGS" -o "./${{ steps.values.outputs.binary-name }}" .
+          LDFLAGS: -s -w -X github.com/tarampampam/mikrotik-hosts-parser/internal/pkg/version.version=${{ steps.values.outputs.version }}
+        run: go build -trimpath -ldflags "$LDFLAGS" -o "./${{ steps.values.outputs.binary-name }}" ./cmd/mikrotik-hosts-parser/
 
       - name: Upload artifact
         uses: actions/upload-artifact@v2
