+++ cmd/mikrotik-hosts-parser/main_test.go
