+++ internal/pkg/cmd/version/command_test.go
