+++ pkg/mikrotik/dns/static_entry.go
