+++ test/testdata/hosts/spy.txt
